Create Signup component
